+++ src/core/CommunicationBackend.cpp
@@ -4,10 +4,7 @@
 #include "core/InputSource.hpp"
 #include "core/state.hpp"
 
-CommunicationBackend::CommunicationBackend(
-    InputSource **input_sources,
-    uint8_t input_source_count
-) {
+CommunicationBackend::CommunicationBackend(InputSource **input_sources, size_t input_source_count) {
     _gamemode = nullptr;
     _input_sources = input_sources;
     _input_source_count = input_source_count;
@@ -18,13 +15,13 @@ InputState &CommunicationBackend::GetInputs() {
 }
 
 void CommunicationBackend::ScanInputs() {
-    for (int i = 0; i < _input_source_count; i++) {
+    for (size_t i = 0; i < _input_source_count; i++) {
         _input_sources[i]->UpdateInputs(_inputs);
     }
 }
 
 void CommunicationBackend::ScanInputs(InputScanSpeed input_source_filter) {
-    for (int i = 0; i < _input_source_count; i++) {
+    for (size_t i = 0; i < _input_source_count; i++) {
         InputSource *input_source = _input_sources[i];
         if (input_source->ScanSpeed() == input_source_filter) {
             input_source->UpdateInputs(_inputs);
